requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
astroid==1.6.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
cryptography==1.7.2
dash==0.21.1
decorator==4.3.0
enum34==1.1.6
ez-setup==0.9
flasgger==0.8.3
Flask==1.0.2
Flask-Compress==1.4.0
Flask-Cors==3.0.6
Flask-JWT==0.3.2
Flask-Migrate==2.1.1
flask-paginate==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
funcsigs==1.0.2
gunicorn==19.8.1
idna==2.6
ipaddress==1.0.22
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
nbformat==4.4.0
nose==1.3.7
nose2==0.7.4
pbr==4.0.2
plotly==2.6.0
psycopg2==2.7.4
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.4
pylint==1.9.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
traitlets==4.3.2
typing==3.5.3.0
urllib3==1.22
validate-email==1.3
Werkzeug==0.14.1
wrapt==1.10.11
